+++ R/get_predicted.R
@@ -48,10 +48,29 @@
 #'   will fix the value of the smooth to its average, so that the predictions
 #'   are not depending on it. (default), `mean()`, or
 #'   `bayestestR::map_estimate()`.
+#' @param ci The interval level (default `0.95`, i.e., `95%` CI).
+#' @param ci_type Can be `"prediction"` or `"confidence"`. Prediction
+#'   intervals show the range that likely contains the value of a new
+#'   observation (in what range it would fall), whereas confidence intervals
+#'   reflect the uncertainty around the estimated parameters (and gives the
+#'   range of the link; for instance of the regression line in a linear
+#'   regressions). Prediction intervals account for both the uncertainty in the
+#'   model's parameters, plus the random variation of the individual values.
+#'   Thus, prediction intervals are always wider than confidence intervals.
+#'   Moreover, prediction intervals will not necessarily become narrower as the
+#'   sample size increases (as they do not reflect only the quality of the fit).
+#'   This applies mostly for "simple" linear models (like `lm`), as for
+#'   other models (e.g., `glm`), prediction intervals are somewhat useless
+#'   (for instance, for a binomial model for which the dependent variable is a
+#'   vector of 1s and 0s, the prediction interval is... `[0, 1]`).
+#' @param ci_method The method for computing p values and confidence intervals. Possible values depend on model type.
+#'   + `NULL` uses the default method, which varies based on the model type.
+#'   + Most frequentist models: `"normal"` (default).
+#'   + Bayesian models: "quantile" (default), "hdi", "eti".
+#'   + Mixed effects **lme4** models: `"normal"` (default), `"satterthwaite"`, `"kenward"`.
+#' @param dispersion_method Bootstrap dispersion and Bayesian posterior summary: `"sd"` or `"mad"`.
 #' @param ... Other argument to be passed, for instance to `get_predicted_ci()`.
-#'   This can be used to request confidence intervals based on robust standard
-#'   errors, e.g. by specifying the `vcov` and `vcov_args` arguments from
-#'   `get_predicted_ci()` directly in the call to `get_predicted()`.
+#' @inheritParams get_varcov
 #' @inheritParams get_df
 #'
 #' @return The fitted values (i.e. predictions for the response). For Bayesian
@@ -163,8 +182,19 @@ get_predicted <- function(x, ...) {
 
 # default methods ---------------------------
 
+#' @rdname get_predicted
 #' @export
-get_predicted.default <- function(x, data = NULL, predict = NULL, verbose = TRUE, ...) {
+get_predicted.default <- function(x,
+                                  data = NULL,
+                                  predict = "expectation",
+                                  ci = 0.95,
+                                  ci_type = "confidence",
+                                  ci_method = NULL,
+                                  dispersion_method = "sd",
+                                  vcov = NULL,
+                                  vcov_args = NULL,
+                                  verbose = TRUE,
+                                  ...) {
 
   # evaluate arguments
   args <- .get_predicted_args(x, data = data, predict = predict, verbose = verbose, ...)
@@ -179,6 +209,12 @@ get_predicted.default <- function(x, data = NULL, predict = NULL, verbose = TRUE
   predict_args <- compact_list(list(x, newdata = args$data, type = args$type, dot_args))
   predictions <- tryCatch(do.call("predict", predict_args), error = function(e) NULL)
 
+  # may fail due to invalid "dot_args", so try shorter argument list
+  if (is.null(predictions)) {
+    predictions <- tryCatch(do.call("predict", compact_list(list(x, newdata = args$data, type = args$type))), error = function(e) NULL)
+  }
+
+  # still fails? try fitted()
   if (is.null(predictions)) {
     predictions <- tryCatch(do.call("fitted", predict_args), error = function(e) NULL)
   }
@@ -191,6 +227,9 @@ get_predicted.default <- function(x, data = NULL, predict = NULL, verbose = TRUE
         predictions,
         data = args$data,
         ci_type = args$ci_type,
+        ci_method = ci_method,
+        vcov = vcov,
+        vcov_args = vcov_args,
         ...
       )
     },
@@ -476,10 +515,12 @@ get_predicted.afex_aov <- function(x, data = NULL, ...) {
       # Delta method; SE * deriv( inverse_link(x) wrt lin_pred(x) )
       mu_eta <- tryCatch(abs(get_family(x)$mu.eta(predictions)), error = function(e) NULL)
       if (is.null(mu_eta)) {
+        ci_data[se_col] <- NULL
         if (isTRUE(verbose)) {
           warning(format_message(
-            "Could not apply Delta method to transform standard errors.",
-            "These are returned on the link-scale."
+            'Could not apply Delta method to transform standard errors.',
+            'You may be able to obtain standard errors by using the ',
+            '`predict="link"` argument value.'
           ), call. = FALSE)
         }
       } else {