+++ R/get_predicted_se.R
@@ -5,6 +5,7 @@ get_predicted_se <- function(x,
                              ci_type = "confidence",
                              vcov = NULL,
                              vcov_args = NULL,
+                             ci_method = NULL,
                              ...) {
 
   # Matrix-multiply X by the parameter vector B to get the predictions, then
@@ -13,12 +14,27 @@ get_predicted_se <- function(x,
   # the diagonal of this matrix represent the standard errors of the predictions,
   # which are then multiplied by 1.96 for the confidence intervals.
 
-  vcovmat <- .get_predicted_ci_vcov(
-    x,
-    vcov_fun = vcov,
-    vcov_args = vcov_args,
-    ...
-  )
+
+  # kenward-roger adjusts both the dof and the varcov
+  if (isTRUE(ci_method %in% c("kenward-roger", "kenward"))) {
+    if (is.null(vcov) && is.null(vcov_args)) {
+      check_if_installed("pbkrtest")
+      vcovmat <- as.matrix(pbkrtest::vcovAdj(x))
+    } else {
+      msg <- 'The `vcov` argument cannot be used together with the `ci_method="kenward-roger"` argument.'
+      stop(format_message(msg))
+    }
+
+  # all other varcov types can be supplied manually
+  } else {
+    vcovmat <- get_varcov(
+      x,
+      vcov = vcov,
+      vcov_args = vcov_args,
+      ...
+    )
+  }
+
   mm <- .get_predicted_ci_modelmatrix(x, data = data, vcovmat = vcovmat)
 
   # return NULL for fail
@@ -81,139 +97,18 @@ get_predicted_se <- function(x,
 
 
 
-# Get Variance-covariance Matrix ---------------------------------------------------
-
-.get_predicted_ci_vcov <- function(x,
-                                   vcov_fun = NULL,
-                                   vcov_args = NULL,
-                                   verbose = TRUE,
-                                   ...) {
-  dots <- list(...)
-
-  # deprecated
-  if (isTRUE(verbose) && "vcov_type" %in% names(dots)) {
-    warning(format_message("The `vcov_type` argument is superseded by the `vcov_args` argument."), call. = FALSE)
-  }
-  if (isTRUE(verbose) && "robust" %in% names(dots)) {
-    warning(format_message("The `robust` argument is superseded by the `vcov` argument."), call. = FALSE)
-  }
-
-  if (is.null(vcov_args)) {
-    vcov_args <- list()
-  }
-
-  # deprecated: `vcov_estimation`
-  if (is.null(vcov_fun) && "vcov_estimation" %in% names(dots)) {
-    vcov_fun <- dots[["vcov_estimation"]]
-  }
-
-  # deprecated: `robust`
-  if (isTRUE(dots[["robust"]]) && is.null(vcov_fun)) {
-    dots[["robust"]] <- NULL
-    vcov_fun <- "HC3"
-  }
-
-  # deprecated: `vcov_type`
-  if ("vcov_type" %in% names(dots)) {
-    if (!"type" %in% names(vcov_args)) {
-      vcov_args[["type"]] <- dots[["vcov_type"]]
-    }
-  }
-
-  # vcov_fun is a matrix
-  if (is.matrix(vcov_fun)) {
-    return(vcov_fun)
-  }
-
-  # vcov_fun is a function
-  if (is.function(vcov_fun)) {
-    if (is.null(vcov_args) || !is.list(vcov_args)) {
-      args <- list(x)
-    } else {
-      args <- c(list(x), vcov_args)
-    }
-    .vcov <- do.call("vcov_fun", args)
-    return(.vcov)
-  }
-
-  # type shortcuts: overwrite only if not supplied explicitly by the user
-  if (!"type" %in% names(vcov_args)) {
-    if (isTRUE(vcov_fun %in% c(
-      "HC0", "HC1", "HC2", "HC3", "HC4", "HC4m", "HC5",
-      "CR0", "CR1", "CR1p", "CR1S", "CR2", "CR3", "xy",
-      "residual", "wild", "mammen", "webb"
-    ))) {
-      vcov_args[["type"]] <- vcov_fun
-    }
-  }
-
-  # default vcov matrix
-  if (is.null(vcov_fun)) {
-    .vcov <- get_varcov(x, ...)
-    return(.vcov)
-  }
-
-  if (!grepl("^(vcov|kernHAC|NeweyWest)", vcov_fun)) {
-    vcov_fun <- switch(vcov_fun,
-      "HC0" = ,
-      "HC1" = ,
-      "HC2" = ,
-      "HC3" = ,
-      "HC4" = ,
-      "HC4m" = ,
-      "HC5" = ,
-      "HC" = "vcovHC",
-      "CR0" = ,
-      "CR1" = ,
-      "CR1p" = ,
-      "CR1S" = ,
-      "CR2" = ,
-      "CR3" = ,
-      "CR" = "vcovCR",
-      "xy" = ,
-      "residual" = ,
-      "wild" = ,
-      "mammen" = ,
-      "webb" = ,
-      "BS" = "vcovBS",
-      "OPG" = "vcovOPG",
-      "HAC" = "vcovHAC",
-      "PC" = "vcovPC",
-      "CL" = "vcovCL",
-      "PL" = "vcovPL"
-    )
-  }
-
-  # check if required package is available
-  if (vcov_fun == "vcovCR") {
-    check_if_installed("clubSandwich", reason = "to get cluster-robust standard errors")
-    fun <- get(vcov_fun, asNamespace("clubSandwich"))
-  } else {
-    check_if_installed("sandwich", reason = "to get robust standard errors")
-    fun <- try(get(vcov_fun, asNamespace("sandwich")), silent = TRUE)
-    if (!is.function(fun)) {
-      stop(sprintf("`%s` is not a function exported by the `sandwich` package.", vcov_fun))
-    }
-  }
-
-  # extract variance-covariance matrix
-  .vcov <- do.call(fun, c(list(x), vcov_args))
-
-  .vcov
-}
-
 
 # Get Model matrix ------------------------------------------------------------
 
 .get_predicted_ci_modelmatrix <- function(x, data = NULL, vcovmat = NULL, ...) {
   resp <- find_response(x)
-  if (is.null(vcovmat)) vcovmat <- .get_predicted_ci_vcov(x, ...)
+  if (is.null(vcovmat)) vcovmat <- get_varcov(x, ...)
 
 
   if (is.null(data)) {
     mm <- get_modelmatrix(x)
   } else {
-    if (!all(resp %in% data)) data[[resp]] <- 0 # fake response
+    if (!all(resp %in% colnames(data))) data[[resp]] <- 0 # fake response
     # else, model.matrix below fails, e.g. for log-terms
     attr(data, "terms") <- NULL